+++ src/c/v2.c
@@ -7,12 +7,12 @@
 uint64_t* v2(int *adjacency, const uint32_t nodes) {
     printf("----------Version 2 is called----------\n");
     uint64_t *vertices = (uint64_t *)calloc(nodes, sizeof(uint64_t));
-    int count = 0;
+    uint64_t count = 0;
 
     struct timespec tic;
     struct timespec toc;
     clock_gettime(CLOCK_MONOTONIC, &tic);
-    printf("Tic: %ld seconds and %ld nanoseconds\n", tic.tv_sec, tic.tv_nsec);
+    printf("Tic: %lu seconds and %lu nanoseconds\n", tic.tv_sec, tic.tv_nsec);
     int i, j, k;
     for (i=0;i<nodes-2;i++){
         for (j=i+1;j<nodes-1;j++) {
@@ -27,10 +27,10 @@ uint64_t* v2(int *adjacency, const uint32_t nodes) {
         }
     }
     clock_gettime(CLOCK_MONOTONIC, &toc);
-    printf("Toc: %ld seconds and %ld nanoseconds\n", toc.tv_sec, toc.tv_nsec);
+    printf("Toc: %lu seconds and %lu nanoseconds\n", toc.tv_sec, toc.tv_nsec);
     double diff = diff_time(tic, toc);
     printf("Time elapsed (seconds): %0.6f\n", diff);
-    printf("Total triangles: %d\n", count);
+    printf("Total triangles: %lu\n", count);
 
     return vertices; // Note: consider to free the allocated memory
 }
